voice_stream_engine/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
websockets==12.0
redis==5.0.1
python-dotenv==1.0.1
pydantic==2.6.3
pydantic-settings==2.2.1
# AI Providers
openai==1.13.3
deepgram-sdk==3.1.0
elevenlabs==0.2.27
# Utils
numpy==1.26.4
scipy==1.12.0
httpx==0.27.0
# Testing
pytest==8.0.2
pytest-asyncio==0.23.5
qdrant-client==1.7.0